Altered copyright notices and licensing.
I've done more looking into how licensing software works and come to the discovery that CCY-BY 4.0 is not recommended for use with software.
As such I've changed the licensing for the project as a whole to the MIT/Expat license, which in my eyes serves the same purpose.
Additionally, license notices in individual files have been replaced with links to the project github page and a copyright notice with the understanding that the MIT license already requires the redistribution of a single file to include the appropriate license notice or be re-licensed to remove that requirement.

Finally, the years on said copyright notices have been altered to be more accurate after further research on my end. As the project was only published on github in 2021 that will now be the starting year in the range, as is conventionally recommended.